+++ src/AssistantHub.Core/Database/Postgresql/Implementations/AssistantSettingsMethods.cs
@@ -52,7 +52,9 @@ namespace AssistantHub.Core.Database.Postgresql.Implementations
             string query =
                 "INSERT INTO assistant_settings " +
                 "(id, assistant_id, temperature, top_p, system_prompt, max_tokens, context_window, " +
-                "model, enable_rag, enable_retrieval_gate, enable_query_rewrite, query_rewrite_prompt, enable_citations, citation_link_mode, collection_id, retrieval_top_k, retrieval_score_threshold, " +
+                "model, enable_rag, enable_retrieval_gate, enable_query_rewrite, query_rewrite_prompt, " +
+                "enable_reranking, reranker_top_k, reranker_score_threshold, rerank_prompt, " +
+                "enable_citations, citation_link_mode, collection_id, retrieval_top_k, retrieval_score_threshold, " +
                 "search_mode, text_weight, fulltext_search_type, fulltext_language, fulltext_normalization, fulltext_minimum_score, " +
                 "retrieval_include_neighbors, " +
                 "inference_endpoint_id, embedding_endpoint_id, title, logo_url, favicon_url, streaming, created_utc, last_update_utc) " +
@@ -69,6 +71,10 @@ namespace AssistantHub.Core.Database.Postgresql.Implementations
                 (assistantSettings.EnableRetrievalGate ? 1 : 0) + ", " +
                 (assistantSettings.EnableQueryRewrite ? 1 : 0) + ", " +
                 _Driver.FormatNullableString(assistantSettings.QueryRewritePrompt) + ", " +
+                (assistantSettings.EnableReranking ? 1 : 0) + ", " +
+                assistantSettings.RerankerTopK + ", " +
+                _Driver.FormatDouble(assistantSettings.RerankerScoreThreshold) + ", " +
+                _Driver.FormatNullableString(assistantSettings.RerankPrompt) + ", " +
                 (assistantSettings.EnableCitations ? 1 : 0) + ", " +
                 _Driver.FormatNullableString(assistantSettings.CitationLinkMode) + ", " +
                 _Driver.FormatNullableString(assistantSettings.CollectionId) + ", " +
@@ -139,6 +145,10 @@ namespace AssistantHub.Core.Database.Postgresql.Implementations
                 "enable_retrieval_gate = " + (assistantSettings.EnableRetrievalGate ? 1 : 0) + ", " +
                 "enable_query_rewrite = " + (assistantSettings.EnableQueryRewrite ? 1 : 0) + ", " +
                 "query_rewrite_prompt = " + _Driver.FormatNullableString(assistantSettings.QueryRewritePrompt) + ", " +
+                "enable_reranking = " + (assistantSettings.EnableReranking ? 1 : 0) + ", " +
+                "reranker_top_k = " + assistantSettings.RerankerTopK + ", " +
+                "reranker_score_threshold = " + _Driver.FormatDouble(assistantSettings.RerankerScoreThreshold) + ", " +
+                "rerank_prompt = " + _Driver.FormatNullableString(assistantSettings.RerankPrompt) + ", " +
                 "enable_citations = " + (assistantSettings.EnableCitations ? 1 : 0) + ", " +
                 "citation_link_mode = " + _Driver.FormatNullableString(assistantSettings.CitationLinkMode) + ", " +
                 "collection_id = " + _Driver.FormatNullableString(assistantSettings.CollectionId) + ", " +